# api/openapi.yaml
openapi: 3.0.3
info:
  title: BlitterServer API
  version: 1.0.0-draft.3
  license:
    name: MIT
    url: https://opensource.org/license/mit
  description: |
    The BlitterAmp middleman service. Apps (BlitterAmp iOS/desktop) talk only to
    this API; BlitterServer owns the music source (Plex in v1), server-side
    transcoding, acquisition (Lidarr), last.fm, and cross-device taste/love
    state.

    **Auth realms:** device endpoints use bearer tokens; `/admin/api/*` uses an
    admin cookie session. Server push is SSE — see `GET /v1/events` for the
    envelope and event catalog.

    **Users are household profiles** (Netflix-style, passwordless, optional
    PIN; created by the admin in the web admin UI). PIN pairing mints a
    **device token** whose only powers are listing profiles and exchanging for
    **profile tokens** (`POST /v1/profile-tokens`). Everything else requires a
    profile token and is implicitly scoped to that profile: playlists,
    ratings, taste, loves, play history, home rails, last.fm connection.
    Shared across profiles: the library, the artifact cache, and acquisition
    (new-release watching is the union of everyone's artist loves).
    Profile switching in-app = using a different cached profile token; no
    admin involvement after the one-time device pairing approval.

    **Conventions:** JSON everywhere; list endpoints paginate
    with `cursor`/`limit` and return `{items, nextCursor}`; errors are a Problem object.

...
servers:
...
security:
  - deviceToken: []
paths:
...
  /v1/status:
    get:
      operationId: getStatus
      tags: [system]
      summary: Server + integration status
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/capabilities:
    get:
      operationId: getCapabilities
      tags: [system]
      summary: Feature flags the apps use to gate UI
      responses:
        "200":
          description: Capabilities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Capabilities" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/me:
    get:
      operationId: getMe
      tags: [system]
      summary: The calling token's identity (device + profile when profile-scoped)
      responses:
        "200":
          description: Identity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ───────────────────────────── pairing ─────────────────────────────
  /v1/pair:
    post:
      operationId: startPairing
      tags: [pairing]
      summary: Begin PIN pairing
      description: |
        Returns a short-lived human-readable code the user approves in the web
        admin. The app then polls `GET /v1/pair/{pairingId}`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceName, deviceType]
              properties:
                deviceName: { type: string, example: "Nathan's iPhone" }
                deviceType: { type: string, enum: [ios, desktop, web, other] }
                appVersion: { type: string }
      responses:
        "201":
          description: Pairing started
          content:
            application/json:
              schema:
                type: object
                required: [pairingId, code, expiresAt]
                properties:
                  pairingId: { type: string }
                  code: { type: string, example: "7GK4PC", description: Show to the user; entered/approved in web admin }
                  expiresAt: { type: string, format: date-time }
        "429": { $ref: "#/components/responses/Problem" }
...
  /v1/pair/claim:
    post:
      operationId: claimPairCode
      tags: [pairing]
      summary: Claim a QR pairing code (one scan = discovery + auth)
      description: |
        The QR shown in the web admin encodes the server's canonical URL plus a
        short-lived (≤15 min) single-use code. Claiming mints a **device
        token** immediately — no admin approval step (showing the QR was the
        approval). The app then lists profiles and mints profile tokens as
        usual.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, deviceName, deviceType]
              properties:
                code: { type: string, example: "7GK4PC" }
                deviceName: { type: string, example: "Nathan's iPhone" }
                deviceType: { type: string, enum: [ios, desktop, web, other] }
                appVersion: { type: string }
      responses:
        "201":
          description: Device paired
          content:
            application/json:
              schema:
                type: object
                required: [token, deviceId]
                properties:
                  token: { type: string, description: Device token — exchange via /v1/profile-tokens }
                  deviceId: { type: string, example: dev_8f2k1q }
        "404": { $ref: "#/components/responses/NotFound" }
        "410": { description: Code expired or already used, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "429": { $ref: "#/components/responses/Problem" }

  # ───────────────────────────── profiles ─────────────────────────────
  /v1/profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: Household profiles (for the in-app profile switcher)
      description: Accepts device OR profile tokens. Profiles are created by the admin in the web admin UI.
      responses:
        "200":
          description: Profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/profile-tokens:
    post:
      operationId: createProfileToken
      tags: [profiles]
      summary: Exchange into a profile (mints a profile-scoped bearer token)
      description: |
        Accepts device OR profile tokens (switching does not require going back
        through the device token). `pin` required only when the target profile
        has one. Apps cache one token per profile; switching profiles = using a
        different cached token, so queued offline requests keep the attribution
        of the profile they were made under.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profileId]
              properties:
                profileId: { type: string }
                pin: { type: string, nullable: true }
      responses:
        "201":
          description: Profile token
          content:
            application/json:
              schema:
                type: object
                required: [token, profile]
                properties:
                  token: { type: string }
                  profile: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: PIN required or wrong, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/lastfm/callback:
    get:
      operationId: completeLastfmAuth
      tags: [profiles]
      summary: Complete browser-based last.fm authorization
      description: Single-use callback bound to the initiating profile by an opaque, 15-minute state value. State is consumed only after successful encrypted connection persistence. Returns minimal non-cacheable HTML and never exposes the session key.
      security: []
      parameters:
        - { name: state, in: query, required: true, schema: { type: string } }
        - { name: token, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Authorization completed
          headers: { Cache-Control: { schema: { type: string } }, Referrer-Policy: { schema: { type: string } }, Content-Security-Policy: { schema: { type: string } } }
          content: { text/html: { schema: { type: string } } }
        "400":
          description: "Invalid, expired, reused, or mismatched attempt"
          headers: { Cache-Control: { schema: { type: string } }, Referrer-Policy: { schema: { type: string } }, Content-Security-Policy: { schema: { type: string } } }
          content: { text/html: { schema: { type: string } } }
...
  /admin/api/profiles:
    get:
      operationId: adminListProfiles
      tags: [admin]
      summary: List household profiles
      security: [{ adminSession: [] }]
      responses:
        "200":
          description: Profiles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: adminCreateProfile
      tags: [admin]
      summary: Create a profile
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                pin: { type: string, nullable: true, description: Optional 4-8 digit PIN }
                avatarColor: { type: string, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /admin/api/settings/server:
    get:
      operationId: adminGetServerSettings
      tags: [admin]
      summary: Server settings (canonical URL)
      security: [{ adminSession: [] }]
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerSettings" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: adminSetServerSettings
      tags: [admin]
      summary: Update server settings
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServerSettings" }
      responses:
        "204": { description: Saved }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: Per-device token minted via PIN pairing
    adminSession:
      type: apiKey
      in: cookie
      name: blitter_admin
...
  responses:
    Unauthorized:
      description: Missing/invalid credentials
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      required: [title, status]
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string, nullable: true }
        code: { type: string, nullable: true, description: "Machine-readable, e.g. source_unavailable, artifact_not_ready" }
...
    ServerStatus:
      type: object
      required: [version, setupComplete, source]
      properties:
        version: { type: string }
        setupComplete: { type: boolean }
        source:
          type: object
          required: [kind, connected]
          properties:
            kind: { type: string, enum: [none, filesystem, plex] }
            connected: { type: boolean }
            libraryName: { type: string, nullable: true }
        integrations:
          type: object
          properties:
            lidarr: { type: string, enum: [absent, configured, error] }
            lastfm: { type: string, enum: [absent, configured, connected, error] }
    Capabilities:
      type: object
      required: [acquisition, lastfm, transcodeFormats]
      properties:
        acquisition: { type: boolean, description: Love(artist) will acquire; Activity view available }
        lastfm: { type: boolean, description: Instance credentials present — profiles can connect via /v1/me/lastfm }
        transcodeFormats:
          type: array
          items: { type: string, enum: [original, aac] }
          description: aac absent when ffmpeg unavailable
        radio: { type: boolean }
        discovery: { type: boolean, description: "A recommendation provider has operational instance credentials; personalized results additionally require a connected profile" }
    Device:
      type: object
      required: [deviceId, name, type, pairedAt]
      properties:
        deviceId: { type: string, example: dev_8f2k1q }
        name: { type: string }
        type: { type: string, enum: [ios, desktop, web, other] }
        pairedAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time, nullable: true }
    Profile:
      type: object
      required: [profileId, name, hasPin]
      properties:
        profileId: { type: string, example: prf_1a2b3c }
        name: { type: string }
        hasPin: { type: boolean }
        avatarColor: { type: string, nullable: true, description: Hex color for the profile chip }
    Me:
      type: object
      required: [device]
      properties:
        device: { $ref: "#/components/schemas/Device" }
        profile:
          allOf: [{ $ref: "#/components/schemas/Profile" }]
          description: Absent when called with a device token
...
    ServerSettings:
      type: object
      required: [canonicalUrl]
      properties:
        canonicalUrl:
          type: string
          nullable: true
          format: uri
          example: "https://music.example.net"
          description: Public base URL embedded in pairing QRs — set during first-run setup; never a raw LAN IP unless devices genuinely reach it there
...